pip-requirements.txt
Django==1.8.2
django-debug-toolbar==1.3.0
django-extensions==1.5.5
django-relativefilepathfield==1.0.2
docutils==0.12
gunicorn==19.3.0
MySQL-python==1.2.5
pygraphviz==1.2
six==1.9.0
sqlparse==0.1.15
